Organize ApplicationView subcomponents and their tests
