+++ src/features/applicationView/similarCustomersTable/tableTools/customersTableTools.module.scss
